src: warnings fixed up everywhere, and a heap of dead code removed too

+++ src/modules/system/fat/FatFilesystem.cc
@@ -311,7 +311,7 @@ uint64_t FatFilesystem::read(File *pFile, uint64_t location, uint64_t size, uint
         return 0; // can't do it
 
     // validity checking
-    if (static_cast<size_t>(location) >= pFile->getSize())
+    if (location >= pFile->getSize())
     {
         WARNING("FAT: Attempting to read past the EOF [loc=" << location << ", sz=" << size << ", fsz=" << pFile->getSize() << "]");
         return 0;
@@ -319,7 +319,7 @@ uint64_t FatFilesystem::read(File *pFile, uint64_t location, uint64_t size, uint
 
     uint64_t endOffset = location + size;
     uint64_t finalSize = size;
-    if (static_cast<size_t>(endOffset) > pFile->getSize())
+    if (endOffset > pFile->getSize())
     {
         finalSize = pFile->getSize() - location;
